Move get_cache to Cache class
